+++ src/interfaces/IFurnitureLoader.ts
@@ -1,5 +1,10 @@
 import { LoadFurniResult } from "../objects/furniture/util/loadFurni";
+import { FurnitureId } from "./IFurnitureData";
 
 export interface IFurnitureLoader {
-  loadFurni(type: string): Promise<LoadFurniResult>;
+  loadFurni(type: FurnitureFetch): Promise<LoadFurniResult>;
 }
+
+export type FurnitureFetch =
+  | { kind: "id"; id: FurnitureId }
+  | { kind: "type"; type: string };