+++ src/objects/furniture/IFurniture.ts
@@ -1,12 +1,15 @@
+import { FurnitureId } from "../../interfaces/IFurnitureData";
 import { IFurnitureEventHandlers } from "./util/IFurnitureEventHandlers";
 
 export interface IFurniture extends IFurnitureEventHandlers {
-  type: string;
+  id: FurnitureId | undefined;
+  type: string | undefined;
   roomX: number;
   roomY: number;
   roomZ: number;
   direction: number;
   animation: string | undefined;
+  highlight: boolean | undefined;
 }
 
 export type IFurnitureBehavior<T extends IFurniture = IFurniture> = {