+++ src/objects/furniture/WallFurniture.tsx
@@ -6,28 +6,35 @@ import { BaseFurniture } from "./BaseFurniture";
 import { IFurniture, IFurnitureBehavior } from "./IFurniture";
 import { HitEvent } from "../../interfaces/IHitDetection";
 import { getMaskId } from "../room/util/getMaskId";
+import { FurnitureFetchInfo } from "./FurnitureFetchInfo";
+import { getFurnitureFetch } from "./util/getFurnitureFetch";
+import { FurnitureId } from "../../interfaces/IFurnitureData";
 
 export class WallFurniture extends RoomObject implements IFurniture {
   private _baseFurniture: BaseFurniture;
-  private _type: string;
+  private readonly _type: string | undefined;
+  private readonly _id: FurnitureId | undefined;
   private _roomX: number;
   private _roomY: number;
   private _roomZ: number;
   private _animation: string | undefined;
   private _direction: number;
+  private _highlight: boolean = false;
 
-  constructor(options: {
-    roomX: number;
-    roomZ: number;
-    roomY: number;
-    type: string;
-    direction: number;
-    animation?: string;
-    behaviors?: IFurnitureBehavior<WallFurniture>[];
-  }) {
+  constructor(
+    options: {
+      roomX: number;
+      roomZ: number;
+      roomY: number;
+      direction: number;
+      animation?: string;
+      behaviors?: IFurnitureBehavior<WallFurniture>[];
+    } & FurnitureFetchInfo
+  ) {
     super();
 
     this._type = options.type;
+    this._id = options.id;
 
     this._roomX = options.roomX;
     this._roomY = options.roomY;
@@ -36,13 +43,25 @@ export class WallFurniture extends RoomObject implements IFurniture {
     this._direction = options.direction;
 
     this._baseFurniture = new BaseFurniture(
-      options.type,
+      getFurnitureFetch(options),
       options.direction,
       options.animation,
       (direction) => getMaskId(direction, this.roomX, this.roomY)
     );
 
     options.behaviors?.forEach((behavior) => behavior.setParent(this));
+  }
+
+  public get id() {
+    return this._id;
+  }
+
+  public get highlight() {
+    return this._highlight;
+  }
+
+  public set highlight(value) {
+    this._highlight = this._highlight;
   }
 
   public get type() {
@@ -118,6 +137,10 @@ export class WallFurniture extends RoomObject implements IFurniture {
     this._baseFurniture.direction = this.direction;
   }
 
+  private updateHighlight() {
+    this._baseFurniture.highlight = this.highlight;
+  }
+
   private getOffsets(direction: number) {
     if (direction === 2 || direction === 6) return { x: -16, y: -64 };
     if (direction === 4 || direction === 0) return { x: 16, y: -64 };
@@ -133,12 +156,7 @@ export class WallFurniture extends RoomObject implements IFurniture {
     const offsets = this.getOffsets(this.direction);
     if (offsets == null) return;
 
-    const base = this.geometry.getPosition(
-      this.roomX,
-      this.roomY,
-      this.roomZ,
-      "object"
-    );
+    const base = this.geometry.getPosition(this.roomX, this.roomY, this.roomZ);
 
     this._baseFurniture.x = base.x + offsets.x;
     this._baseFurniture.y = base.y + offsets.y;