+++ src/util/parseTileMap.ts
@@ -30,27 +30,10 @@ export function parseTileMap(
   positionOffsets: { x: number; y: number };
   maskOffsets: { x: number; y: number };
 } {
-  const startIndexColumn = findFirstNonEmptyColumnIndex(tilemap);
-  const startIndexRow = findFirstNonEmptyRowIndex(tilemap);
+  assertTileMapHasPadding(tilemap);
 
-  if (startIndexColumn === -1) throw new Error("No column start found");
-  if (startIndexRow === -1) throw new Error("No row start found");
-
-  const endIndexColumn = findLastColumnIndex(tilemap);
-  const endIndexRow = findLastRowIndex(tilemap);
-
-  const cutoutTilemap = cutoutArray(
-    tilemap,
-    startIndexColumn,
-    startIndexRow,
-    endIndexColumn + 1,
-    endIndexRow + 1
-  );
-
-  const result = initialize2DArray<ParsedTileType>(
-    endIndexColumn - startIndexColumn + 1 + 1,
-    endIndexRow - startIndexRow + 1 + 1,
-    { type: "hidden" }
+  const result: ParsedTileType[][] = tilemap.map((row) =>
+    row.map(() => ({ type: "hidden" as const }))
   );
 
   const rowWallMinimum = new RowWall();
@@ -72,22 +55,22 @@ export function parseTileMap(
     }
   }
 
-  for (let y = 0; y < cutoutTilemap.length; y++) {
+  for (let y = 0; y < tilemap.length; y++) {
     let rowWallStartX = -1;
 
-    for (let x = 0; x < cutoutTilemap[y].length; x++) {
-      const resultX = x + 1;
-      const resultY = y + 1;
+    for (let x = 0; x < tilemap[y].length; x++) {
+      const resultX = x;
+      const resultY = y;
 
       const wallPositionX = resultX - 1;
       const wallPositionY = resultY - 1;
 
-      const tileInfo = getTileInfo(cutoutTilemap, x, y);
+      const tileInfo = getTileInfo(tilemap, x, y);
 
       if (!tileInfo.rowDoor || hasDoor) {
         const rowWallAllowed = rowWallMinimum.isWallAllowed(wallPositionX);
         if (tileInfo.rowEdge && tileInfo.height != null && rowWallAllowed) {
-          const belowTileInfo = getTileInfo(cutoutTilemap, x - 1, y + 1);
+          const belowTileInfo = getTileInfo(tilemap, x - 1, y + 1);
 
           result[resultY][wallPositionX] = {
             type: "wall",
@@ -179,7 +162,6 @@ export function parseTileMap(
     x: 1,
     y: 1,
   };
-  const doorOffset = hasDoor ? 1 : 0;
 
   return {
     tilemap: result,
@@ -188,14 +170,8 @@ export function parseTileMap(
     // When the tilemap has a door, we offset the objects in the room by one in the x direction.
     // This makes it so objects appear at the same position, for a room without a door
     // and for a room with a door.
-    positionOffsets: {
-      x: wallOffsets.x + doorOffset,
-      y: wallOffsets.y,
-    },
-    maskOffsets: {
-      x: doorOffset,
-      y: 0,
-    },
+    positionOffsets: { x: 0, y: 0 },
+    maskOffsets: { x: -wallOffsets.x, y: -wallOffsets.y },
   };
 }
 
@@ -213,93 +189,34 @@ class RowWall {
   }
 }
 
-function cutoutArray(
-  tilemap: TileType[][],
-  startX: number,
-  startY: number,
-  endX: number,
-  endY: number
-) {
-  const width = endX - startX;
-  const height = endY - startY;
+function assertTileMapHasPadding(tilemap: TileType[][]) {
+  if (tilemap.length < 1) throw new Error("Tilemap has no rows");
 
-  const result = initialize2DArray<TileType>(width, height, "0");
+  let doorCount = 0;
 
-  for (let y = 0; y < height; y++) {
-    for (let x = 0; x < width; x++) {
-      result[y][x] = tilemap[y + startY][x + startX];
-    }
-  }
-
-  return result;
-}
-
-function initialize2DArray<T, TD = T>(
-  maxX: number,
-  maxY: number,
-  defaultValue: TD | T
-) {
-  const res: (T | TD)[][] = [];
-
-  for (let y = 0; y < maxY; y++) {
-    const row: (T | TD)[] = [];
-    for (let x = 0; x < maxX; x++) {
-      row.push(defaultValue);
-    }
-    res.push(row);
-  }
-
-  return res;
-}
-
-function findFirstNonEmptyColumnIndex(tilemap: TileType[][]) {
-  for (let x = 0; x < tilemap[0].length; x++) {
-    for (let y = 0; y < tilemap.length; y++) {
-      if (tilemap[y][x] !== "x") return x;
-    }
-  }
-
-  return -1;
-}
-
-function findFirstNonEmptyRowIndex(tilemap: TileType[][]) {
   for (let y = 0; y < tilemap.length; y++) {
-    for (let x = 0; x < tilemap[y].length; x++) {
-      if (tilemap[y][x] !== "x") return y;
-    }
+    const row = tilemap[y];
+
+    if (row.length < 1) throw new Error("Tilemap row was empty.");
+    if (doorCount > 0 && row[0] !== "x")
+      throw new Error(paddingErrorMessage(`row ${y}`));
+
+    doorCount++;
   }
 
-  return -1;
-}
-
-function findLastColumnIndex(tilemap: TileType[][]) {
-  let lastColumn = 0;
   for (let x = 0; x < tilemap[0].length; x++) {
-    let hasTile = false;
-    for (let y = 0; y < tilemap.length; y++) {
-      hasTile = hasTile || isTile(getNumberOfTileType(tilemap[y][x]));
-    }
+    const cell = tilemap[0][x];
 
-    if (hasTile) {
-      lastColumn = x;
-    }
+    if (cell !== "x") throw new Error(paddingErrorMessage(`column ${x}`));
   }
-
-  return lastColumn;
 }
 
-function findLastRowIndex(tilemap: TileType[][]) {
-  let lastRow = 0;
-  for (let y = 0; y < tilemap.length; y++) {
-    let hasTile = false;
-    for (let x = 0; x < tilemap.length; x++) {
-      hasTile = hasTile || isTile(getNumberOfTileType(tilemap[y][x]));
-    }
+const paddingErrorMessage = (text: string) => `No padding for ${text}
+- There should be one 'x' at the beginning of each row
+- There should be a full row of 'x' as the first row
 
-    if (hasTile) {
-      lastRow = y;
-    }
-  }
-
-  return lastRow;
-}
+Please ensure that the tilemap is padded like the following example:
+xxx
+xoo
+xoo
+`;